requirements.txt
django>=4.2
djangorestframework
pymysql
cryptography
python-dotenv
openai
django-cors-headers
